:recycle: refactor : 응답 객체와 영속화 객체에 변환을 도메인 클래스에서 분리하고 별도의 mapper 파일로 변경
